Group commands based on side they operate on
Grouped together commands based on the sides they are mostly concern of
(handoff vs endpoint) and renamed modules accordingly.
