+++ dbb_buffer_mngr/__init__.py
@@ -1,4 +1,3 @@
-from dbb_buffer_mngr.cleaner import Cleaner
-from dbb_buffer_mngr.porter import Porter, Wiper
-from dbb_buffer_mngr.scanner import Scanner
+from dbb_buffer_mngr.endpoint import Porter, Wiper
+from dbb_buffer_mngr.handoff import Scanner, Cleaner
 from dbb_buffer_mngr.schema import SCHEMA
