+++ dbb_buffer_mngr/endpoint.py
